# Part of #1404: Avoid flushing aggregators in lockstep